Move Stack.Build.Coverage to Stack.Coverage
